+++ login.php
@@ -30,6 +30,7 @@ if (rssg_is_logged_in()) {
 $lang = [];
 $lang_formular = [];
 $lang_navigation_top = [];
+$lang_navigation_top['zugang'] = rssg_nav_zugang();
 $lang_formular['meldung'] = '';
 $username = '';